requirements.txt
# fastapi
fastapi
uvicorn
tortoise-orm

#for database
sqlalchemy
psycopg2

#for loading environment variables  #new
python-dotenv


#for email validation            #new
pydantic[email]


#hashing              #new
passlib[bcrypt]


#for jwt tokens
python-jose


#for accessing from data
python-multipart